Make it more `print` like to easy change prints
Print function support a spelling like that:
    print(‘some’, ‘thing’)
to print ‘some thing’.  To make that red e.g. one has to concatenate
the strings first like that:
    print(red(‘some thing’))

With that it is now possible to easily write it like that:
    print(red(‘some’, ‘thing’))

It also supports to change the separator then, just like the `print`
function.